Split the actual music services into two separate parts.
Created a DataLoader that will run on initialisation of the server, if any files are found within the data folder, it will be parsed based off of the format that has been supplied and then move that file to the archive folder.

searchSingers method needs to be worked on.

Once this has been completed, the front end can be started.